requirements.txt
Flask==2.3.2
flask-cors==4.0.0
gunicorn
python-dotenv
requests
huggingface_hub
pillow
google-genai